requirements.txt
flask
flask-cors
pdfplumber
python-docx
openai
